% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_dataset.R
\name{acquire_dataset}
\alias{acquire_dataset}
\title{acquire_dataset}
\usage{
acquire_dataset(DatasetID)
}
\arguments{
\item{DatasetID}{A string.}
}
\description{
\code{acquire_dataset} purchases and downloads a dataset from the hut.
}
\details{
This function allows the user to purchase a dataset by specifying its DatasetID.
}
\examples{
acquire_dataset ("shdurr8nf80SLK")
acquire_dataset (DatasetID = "shdurr8nf80SLK")

}
