% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_available_datasets.R
\name{browse_available_datasets}
\alias{browse_available_datasets}
\title{browse_available_datasets}
\usage{
browse_available_datasets()
}
\value{
The metadata of all available data is returned as a dataframe.
}
\description{
\code{browse_available_datasets} provides information about all datasets available to the user.
}
\details{
This function calls the hut and returns a dataframe of all available data that the user has not
  yet acquired.
}
\examples{
browse_available_datasets()

}
