% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_connection.R
\name{check_connection}
\alias{check_connection}
\title{check_connection}
\usage{
check_connection()
}
\value{
The two tokens are assigned as global variables.
}
\description{
\code{check_connection} sends a post request to test whether connections has been established
}
\details{
Sends a post request to the /alive endpoint.
  Provides a message in the Console including the Status Code and whether the connection was successful.
}
\examples{
check_connection()

}
