% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hut_url_valid.R
\name{check_hut_url_valid}
\alias{check_hut_url_valid}
\title{check_hut_url_valid}
\usage{
check_hut_url_valid(urlToCheck)
}
\arguments{
\item{urlToCheck}{A string}
}
\value{
Boolean
}
\description{
\code{check_hut_url_valid} checks if the URL entered is valid
}
\details{
Checks if the URL entered by the user is of the form https://XXXXXXX.web.app/
}
\examples{
check_hut_url_valid("https://XXXXXXX.web.app/")

}
