% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url_and_tokens.R
\name{check_url_and_tokens}
\alias{check_url_and_tokens}
\title{check_url_and_tokens}
\usage{
check_url_and_tokens()
}
\value{
Boolean
}
\description{
\code{check_url_and_tokens} checks if the URL and tokens are valid
}
\details{
Checks if the Hut URL exists in the global environment and if it is of the form https://XXXXXXX.web.app/
  Then checks if the Access and ID tokens exist.
  Returns TRUE if all conditions are met.
}
\examples{
check_url_and_tokens()

}
