% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_object.R
\name{download_data_object}
\alias{download_data_object}
\title{download_data_object}
\usage{
download_data_object(DatasetID)
}
\arguments{
\item{DatasetID}{A string}
}
\value{
Returns a list
}
\description{
\code{download_data_object} downloads a data object
}
\details{
This function downloads a JSON data object and converts from JSON
}
\examples{
extract_dataframe_from_json(myJsonObject)

}
