% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{download_dataset}
\usage{
download_dataset(DatasetNumber, Endpoint)
}
\arguments{
\item{DatasetNumber}{An integer.}

\item{Endpoint}{A string}
}
\value{
Returns a dataframe.
}
\description{
\code{download_dataset} downloads a dataset.
}
\details{
This function allows the user to download a dataset by specifying its number in the dataframe of available datsets.
 User must also specify the endpoint.
}
\examples{
download_dataset (2, "/user/getDatasets")
download_dataset (DatasetNumber = 3, Endpoint = "/user/getDatasets")
}
