% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_uploaded_dataset.R
\name{download_uploaded_dataset}
\alias{download_uploaded_dataset}
\title{download_uploaded_dataset}
\usage{
download_uploaded_dataset(DatasetNumber)
}
\arguments{
\item{DatasetNumber}{An integer.}
}
\value{
Returns a dataframe.
}
\description{
\code{download_uploaded_dataset} downloads a dataset that the user has themselves uploaded.
}
\details{
This function allows the user to download a dataset by specifying its number in the dataframe of available datsets.
}
\examples{
download_uploaded_dataset (2)
download_uploaded_dataset (DatasetNumber = 3)
}
