% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dataframe_from_json.R
\name{extract_dataframe_from_json}
\alias{extract_dataframe_from_json}
\title{extract_dataframe_from_json}
\usage{
extract_dataframe_from_json(JsonObject)
}
\arguments{
\item{JsonObject}{A JSON object}
}
\value{
Returns a dataframe
}
\description{
\code{extract_dataframe_from_json} parses JSON object to retrieve dataframe
}
\details{
This function extracts the dataframe from the JSON object that is retrieved from the hut.
}
\examples{
extract_dataframe_from_json(myJsonObject)

}
