% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_datasets.R
\name{get_all_datasets}
\alias{get_all_datasets}
\title{get_all_datasets}
\usage{
get_all_datasets()
}
\value{
The metadata of all uploaded or purchased data is returned as a dataframe.
}
\description{
\code{get_all_datasets} provides information about all datasets immediately available to the user.
}
\details{
This function calls the hut and returns a dataframe of all data that the current user
  has either purchased or has uploaded themselves.
}
\examples{
get_all_datasets()

}
