% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasetID.R
\name{get_datasetID}
\alias{get_datasetID}
\title{get_datasetID}
\usage{
get_datasetID(DatasetNumber, Endpoint)
}
\arguments{
\item{DatasetNumber}{An integer.}

\item{Endpoint}{A string}
}
\value{
Returns a string
}
\description{
\code{get_datasetID} retrieves dataset ID
}
\details{
Dataset ID is not currently specified in the table of datasets so requires a separate post request.
 User must also specify the endpoint.
}
\examples{
get_datasetID(DatasetNumber=DatasetNumber, Endpoint=Endpoint)

}
