% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets}
\usage{
get_datasets(Endpoint)
}
\arguments{
\item{Endpoint}{A string}
}
\value{
The metadata of relevant data is returned as a dataframe.
}
\description{
\code{get_datasets} provides information about datasets available to the user.
}
\details{
This function calls the hut and returns a dataframe of available data.
  The subset of data displayed depends on the endpoint specified.
}
\examples{
get_datasets(Endpoint = "/user/getDatasets")
get_datasets("/user/getUploadedDatasets")

}
