% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hut_url.R
\name{get_hut_url}
\alias{get_hut_url}
\title{get_hut_url}
\usage{
get_hut_url()
}
\value{
The two URLs are assigned as global variables.
}
\description{
\code{get_hut_url} asks the user for the URL of their hut
}
\details{
This function asks for the URL of the data hut, input via the console.
  It also derives a URL for endpoints for pushing and pulling data.
}
\examples{
get_hut_url()

}
