% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tokens.R
\name{get_tokens}
\alias{get_tokens}
\title{get_tokens}
\usage{
get_tokens()
}
\value{
The two tokens are assigned as global variables.
}
\description{
\code{get_tokens} prompts the user to retrieve tokens from the hut website and sets them as global variables
}
\details{
Users are prompted to copy and paste access and ID tokens from the hut website and paste them into the console.
}
\examples{
get_tokens()

}
