% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_hut.R
\name{initiate_hut}
\alias{initiate_hut}
\title{initiate_hut}
\usage{
initiate_hut()
}
\value{
The URL and tokens are set as global variables.
  A console message indicates if the connection was successful.
}
\description{
\code{initiate_hut} establishes a connection between R and Hut34.
}
\details{
This function asks for the URL of the data hut, input via the console.
It then prompts the user to enter access and ID tokens.
Finally it sends a post request to check the connection.
}
\examples{
initiate_hut()

}
