% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_post_request function.R
\name{make_post_request}
\alias{make_post_request}
\title{Make post requests}
\usage{
make_post_request(Endpoint, Body = "EMPTY")
}
\arguments{
\item{Endpoint}{A string.}

\item{Body}{A string.}
}
\value{
The response to the post request is returned as a list.
}
\description{
\code{make_post_request} sends a post request to the hut.
}
\details{
This function sends a post request to the hut after first adding tokens to the body of the post.
  It first adds access and ID tokens to the body of the post.
}
\examples{
make_post_request(Endpoint = "/alive")
make_post_request(Endpoint = "/user/downloadFile", Body = paste0("\"dataSetId\":\"", myDatasetID,"\""))
}
