% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_hut_in_browser.R
\name{open_hut_in_browser}
\alias{open_hut_in_browser}
\title{open_hut_in_browser}
\usage{
open_hut_in_browser()
}
\description{
\code{open_hut_in_browser} opens the hut web interface.
}
\details{
This function switches the user from R to their default browser and takes them to the address they input when
  initiating the hut.
}
\examples{
open_hut_in_browser ()

}
