% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchase_dataset_function.R
\name{purchase_dataset}
\alias{purchase_dataset}
\title{purchase_dataset}
\usage{
purchase_dataset(DatasetNumber, Download = FALSE)
}
\arguments{
\item{DatasetNumber}{An integer.}

\item{Download}{A Boolean}
}
\value{
If Download is TRUE then returns a dataframe.
}
\description{
\code{purchase_dataset} purchases and downloads a dataset from the hut.
}
\details{
This function allows the user to purchase a dataset by specifying its number in the dataframe of available datsets.
  If Download is TRUE then the dataset is instantly downloaded. If FALSE then it can be downloaded at a later time.
}
\examples{
purchase_dataset (2)
purchase_dataset (DatasetNumber = 3, Download = TRUE)
}
