% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dataframe_columns_to_numeric.R
\name{transform_dataframe_columns_to_numeric}
\alias{transform_dataframe_columns_to_numeric}
\title{transform_dataframe_columns_to_numeric}
\usage{
transform_dataframe_columns_to_numeric(Dataframe)
}
\arguments{
\item{Dataframe}{A dataframe}
}
\value{
Returns a dataframe
}
\description{
\code{transform_dataframe_columns_to_numeric} transforms dataframe columns into numberic (rather than strings)
}
\details{
This may not be needed any more but at one point all data retrieved from the hut was treated as a string.
  The function determines if all the data in a column is numeric, then sets the column to numeric.
}
\examples{
transform_dataframe_columns_to_numeric (myDf)

}
