% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_hut.R
\name{upload_dataset}
\alias{upload_dataset}
\title{upload_dataset}
\usage{
upload_dataset(
  Data,
  Name,
  Description,
  Parent = "No parent",
  Custodial = TRUE,
  Price = 3.14,
  Image = "DEFAULT"
)
}
\arguments{
\item{Data}{A dataframe}

\item{Name}{A string}

\item{Description}{A string}

\item{Parent}{A string}

\item{Custodial}{A Boolean}

\item{Price}{A number}

\item{Image}{A string}
}
\value{
Returns the reply to the post request as a list - I think!!!
}
\description{
\code{upload_dataset} uploads a dataset to the hut.
}
\details{
This function adds metadata to a dataframe and uploads it to the hut.
}
\examples{
upload_dataset (Data="myData", Name="Some nice data", Description="This data took ages to collect!", Parent="hfud945r3bdhnsgfte", Custodial=FALSE)
}
