% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_datasets_functions.R
\name{view_all_datasets}
\alias{view_all_datasets}
\title{View all datasets}
\usage{
view_all_datasets()
}
\value{
The metadata of all purchased data is returned as a dataframe.
}
\description{
\code{view_all_datasets} provides information about all datasets available to the user.
}
\details{
This function calls the hut and returns a dataframe of all available data.
  This includes datasets that that the current user has uploaded and datasets
  that are available but have not been purchased.
}
\examples{
view_all_datasets()

}
