% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_datasets.R
\name{view_datasets}
\alias{view_datasets}
\title{view_datasets}
\usage{
view_datasets(Endpoint)
}
\arguments{
\item{Endpoint}{A string}
}
\value{
The metadata of relevant data is returned as a dataframe.
}
\description{
\code{view_datasets} provides information about datasets available to the user.
}
\details{
This function calls the hut and returns a dataframe of available data.
  The subset of data displayed depends on the endpoint specified.
}
\examples{
view_datasets(Endpoint = "/user/getDatasets")
view_datasets("/user/getUploadedDatasets")

}
