% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_my_datasets.R
\name{view_my_datasets}
\alias{view_my_datasets}
\title{view_my_datasets}
\usage{
view_my_datasets()
}
\value{
The metadata of all uploaded or purchased data is returned as a dataframe.
}
\description{
\code{view_my_datasets} provides information about all datasets immediately available to the user.
}
\details{
This function calls the hut and returns a dataframe of all data that the current user
  has either purchased or has uploaded themselves.
}
\examples{
view_my_datasets()

}
