% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_purchased_datasets.R
\name{view_purchased_datasets}
\alias{view_purchased_datasets}
\title{View purchased datasets}
\usage{
view_purchased_datasets()
}
\value{
The metadata of all purchased data is returned as a dataframe.
}
\description{
\code{view_purchased_datasets} provides information about datasets purchased by the user.
}
\details{
This function calls the hut and returns a dataframe of all purchased data
  i.e., data that was uploaded by another user but the current user has
  immediate access to.
}
\examples{
view_purchased_datasets()
}
