% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{download_dataset}
\usage{
download_dataset(DatasetID)
}
\arguments{
\item{DatasetID}{A string}
}
\value{
Returns a dataframe.
}
\description{
\code{download_dataset} downloads a dataset.
}
\details{
This function allows the user to download a dataset by specifying its DatasetID
}
\examples{
df <- download_dataset ("SPPeJC1VInG5VMXN8M6D")

}
