% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_uploaded_datasets.R
\name{view_uploaded_datasets}
\alias{view_uploaded_datasets}
\title{view_uploaded_datasets}
\usage{
view_uploaded_datasets()
}
\value{
The metadata of all uploaded data is returned as a dataframe.
}
\description{
\code{view_uploaded_datasets} provides information about datasets uploaded by the user.
}
\details{
This function calls the hut and returns a dataframe of all uploaded data
  i.e., data that was uploaded by another user but the current user has
  immediate access to.
}
\examples{
view_uploaded_datasets()

}
